removed css files from app.js and put them in correct components, made 2 package.json files, renamed imports on app.js and deleted 2 index files
