+++ client/src/components/addBook/addBook.js
@@ -1,5 +1,6 @@
 import React, {Component} from 'react';
 import DragDrop from './images/x3KMH.jpg'
+import './addBook.css';
 
 class AddBook extends Component {
     constructor(props) {
