--- client/src/components/addBook/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import AddBook from './addBook';
-import Header from '../universal/header';
-import Nav from '../universal/nav';
-import '../universal/universal.css';
-
-function AddBookIndex() {
-    return (
-        <div>
-            <AddBook/>
-        </div>
-
-    )
-}
-
-export default AddBookIndex
