+++ client/src/components/addBook/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "addBook.js"
+}
