+++ client/src/components/app.js
@@ -1,17 +1,16 @@
 import React, {Component} from 'react';
 import Index from './BookInfo/index';
 import Images from './BookInfo/images';
-import LandingPageIndex from './landing/index';
-import AddBookIndex from './addBook/index';
-import BookInfoIndex from './BookInfo/index'
-import './universal/universal.css';
-import './landing/landing.css';
-import './addBook/addBook.css';
-import './BookInfo/book-data.css';
-import {BrowserRouter as Router, Route, Link} from 'react-router-dom';
+import LandingPage from './landing';
+import './landing/package';
+import AddBook from './addBook';
+import './addBook/package'
+import BookInfo from './BookInfo'
 import Nav from './universal/nav';
 import Header from './universal/header';
+import {BrowserRouter as Router, Route, Link} from 'react-router-dom';
 import axios from 'axios';
+
 
 
 
@@ -22,9 +21,9 @@ class App extends Component {
             <Router>
                 <div>
                     <Header/>
-                    <Route exact path={"/"} component={LandingPageIndex}/>
-                    <Route path={"/AddBook"} component={AddBookIndex}/>
-                    <Route path={"/BookInfoIndex"} component={BookInfoIndex}/>
+                    <Route exact path={"/"} component={LandingPage}/>
+                    <Route path={"/AddBook"} component={AddBook}/>
+                    <Route path={"/BookInfoIndex"} component={BookInfo}/>
                     <Nav/>
                 </div>
             </Router>
