+++ client/src/components/landing/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "landing.js"
+}
