+++ client/src/components/universal/nav.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import AddBook from '../addBook/';
-import Search from '../landing/';
+import './universal.css';
+import AddBook from '../addBook';
+import Search from '../landing';
 
 export default () => (
     <footer className="uni-footer">
